# openapi.yaml
openapi: 3.0.1
info:
  title: Kramesh Portfolio Backend API
  description: API for managing portfolio content (education, experience, skills, about, projects, blogs, contact, certifications, auth).
  version: "1.0"
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /skills:
    get:
      tags:
        - skills
      operationId: getSkills
      summary: List all skills
      description: Returns an array of skill records.
      responses:
        200:
          description: Skills found and returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/skills_get_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
    post:
      tags:
        - skills
      operationId: createSkill
      summary: Create a skill
      description: Creates a new skill record. Requires admin authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFields"
            examples:
              create_skill_success:
                $ref: openapi_examples/create_skill_success.json
      responses:
        201:
          description: Skill created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/skills_post_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/skills_post_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
  /skills/{id}:
    put:
      tags:
        - skills
      operationId: updateSkill
      summary: Update a skill by ID
      description: Updates an existing skill record. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d3
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillFields"
            examples:
              update_skill_success:
                $ref: openapi_examples/update_skill_success.json
      responses:
        200:
          description: Skill updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/skills_put_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/skills_put_server_error.json
    delete:
      tags:
        - skills
      operationId: deleteSkill
      summary: Delete a skill by ID
      description: Deletes a skill record. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d3
      responses:
        200:
          description: Skill deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMessage"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/skills_delete_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects:
    get:
      tags:
        - projects
      operationId: getProjects
      summary: List all projects
      description: Returns an array of project records, newest first.
      responses:
        200:
          description: Projects found and returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/projects_get_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
    post:
      tags:
        - projects
      operationId: createProject
      summary: Create a project
      description: Creates a new project record. Requires admin authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFields"
      responses:
        201:
          description: Project created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/projects_post_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/projects_post_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
  /projects/{id}:
    put:
      tags:
        - projects
      operationId: updateProject
      summary: Update a project by ID
      description: Updates an existing project record. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d4
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFields"
      responses:
        200:
          description: Project updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/projects_put_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/projects_put_server_error.json
    delete:
      tags:
        - projects
      operationId: deleteProject
      summary: Delete a project by ID
      description: Deletes a project record. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d4
      responses:
        200:
          description: Project deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMessage"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/projects_delete_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      title: Error response
      type: object
      properties:
        message:
          type: string
          example: Not authorized
      required:
        - message
    DeleteMessage:
      title: Delete confirmation message
      type: object
      properties:
        message:
          type: string
          example: Education deleted
      required:
        - message
...
    SkillFields:
      title: Skill fields
      description: Editable fields of a skill.
      type: object
      properties:
        name:
          type: string
          example: Java
        category:
          type: string
          enum:
            - frontend
            - backend
            - database
            - tools
            - programming
            - other
          example: programming
        icon_url:
          type: string
          example: https://example.com/icons/java.svg
        proficiency:
          type: integer
          example: 80
      required:
        - name
        - category
    Skill:
      title: Skill
      description: A skill entry.
      allOf:
        - $ref: "#/components/schemas/SkillFields"
        - type: object
          properties:
            _id:
              type: string
              pattern: ^[0-9a-fA-F]{24}$
              example: 64f1a2b3c4d5e6f7a8b9c0d3
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            __v:
              type: integer
              example: 0
          required:
            - _id
...
    ProjectFields:
      title: Project fields
      description: Editable fields of a project.
      type: object
      properties:
        title:
          type: string
          example: kramesh-portfolio-backend
        description:
          type: string
          example: Portfolio backend API built with Node.js, Express, and MongoDB.
        tech_stack:
          type: string
          example: Node.js, Express, MongoDB, JWT
        image_url:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/project.jpg
        live_url:
          type: string
          example: https://kramesh.dev
        github_url:
          type: string
          example: https://github.com/KRameshr/kramesh-portfolio-backend
        is_featured:
          type: boolean
          example: true
      required:
        - title
        - description
        - tech_stack
    Project:
      title: Project
      description: A project entry.
      allOf:
        - $ref: "#/components/schemas/ProjectFields"
        - type: object
          properties:
            _id:
              type: string
              pattern: ^[0-9a-fA-F]{24}$
              example: 64f1a2b3c4d5e6f7a8b9c0d4
            image_public_id:
              type: string
              example: project_abc123
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            __v:
              type: integer
              example: 0
          required:
            - _id
...